+++ include/H2DE/H2DE.h
@@ -7,7 +7,7 @@
  * Hydren 2 Dimensional Engine
  * https://github.com/Hydrened/H2DE
  * 
- * Version: 3.0.4
+ * Version: 3.0.5
  * 
  * Copyright (C) 2024-2025
  */