+++ include/H2DE/H2DE_surface.h
@@ -22,6 +22,9 @@ public:
     friend void H2DE_SetSurfaceTextureName(H2DE_Surface* surface, const std::string& textureName);
     friend void H2DE_SetSurfaceColor(H2DE_Surface* surface, const H2DE_ColorRGB& color);
     friend void H2DE_SetSurfaceScaleMode(H2DE_Surface* surface, H2DE_ScaleMode scaleMode);
+    // friend void H2DE_SetSurfaceRotation(H2DE_Surface* surface, float rotation);
+    // friend void H2DE_SetSurfacePivot(H2DE_Surface* surface, const H2DE_LevelPos& pivot);
+    // friend void H2DE_SetSurfaceFlip(H2DE_Surface* surface, H2DE_Flip flip);
 
     friend class H2DE_Engine;
     friend class H2DE_BarObject;